Show train numbers for Israel Railway
A train number is unique per day, and is displayed on the real-time
departure boards in railway stations.

Before this commit, the value of PublishedLineName from SIRI
was displayed for Israel Railways, but it has unclear meaning,
as those are not visible on Israel Railways publications or in the
GTFS feed.