Workaround bad CSVs in the GTFS feed
The CSVs sometimes use " in string fields with no escaping.
This confuses psql if there's an even number of " characters,
because it things these are quotes.

I tried using sed to append \ before them, but that didn't help,
so I decided to cheat and tell psql that the quote character is
U+0007 (BELL) which will probably never appear in the CSV file.