Match town name synonyms to make route names appear cleaner
This avoids the horrifying
"Modi'in Macabim Re'ut Central Station/Alighting-Modi'in-Makabim-Re'ut"

and changes it to a more readable string by removing the
town name from the end.

Previously, the matching was done in a very simplistic way,
which didn't work for Modi'in because they have at least
5 different names to spell that town's full name.

Now using regex, the names look cleaner and the code for adding
more rules is there if I find more similar cases.