Remove load_translations.py and move it to update_feed.sh
load_translations was originally written as a loader for all GTFS
types in the model, but when I realized how slow it was I looked
for faster ways, until I settled on using `psql` and `/copy`.

Even after this was all done with a shell script, translations
were still loaded by the python script because they often contain
quotes, which makes psql's csv parser sad.

With the custom quote character hack this is no longer an issue,
and translations can be loaded with psql directly.

This saves *a lot* of time when loading the GTFS feed. On my
computer (with an SSD) it took 6 minutes to load the feed
(not counting the download time)