Refactor ArrivalGtfsInfo to a function
It's simpler now, and also has better performance (~400ms better!).

Originally I meant for this to be a class so I won't have to use
dictionaries all over the place, and I thought I'd need more
stuff out of it that I won't be searlizing to JSON,
in any case, I'm not entirely sure what the original reasoning was,
and I ended up using the same dictionary in both the renderer
and the JSON-seralized API, and that's obviously better.

So it didn't make much sense to keep it as an object,
which was built, then asynchrounously loaded, then immediately
seralized to JSON.