web ui: Use window.location to figure out the server's address
The web UI code used to assume nobody else is going to run
a curlbus instance, so it was hardcoded to point to
https://curlbus.app

Now that other people started running their own curlbus instances,
the UI should point to whatever server it was retrived from.

A better solution than this would allow the server's URL to be
configurable, but I thought it might be an overkill for this
specific case, so using window.location is good enough for now.

This also fixes a tiny, annoying typo, that apparently everyone
who looked at curlbus and its web UI missed.