GTFS-RT Support for Tel Aviv Shabbat busses!
This is based on work I started back in 2020,
before covid-19 caused the shabbat bus service to be suspended.

Now that the service is back, it's time for curlbus to support it.

It's still a bit buggy with mapping stop_ids between
the MoT GTFS feed and the Tel Aviv GTFS feed,
since not all stops in the Tel Aviv feed have correct stop codes,
and route alternatives is being rendered incorrectly due to Tel Aviv's
GTFS being less eccentric than the MoT one:
Tel Aviv actually uses direction_id properly!
MoT doesn't, and has two route entries per direction. curlbus was built for
the MoT feed, so refactors or kludges will be needed to fix this.

I'll fix the bugs at some point in the future. For now, this is good enough.