Add missing operators
Regular maintenence that can't be avoided. Each operator needs a slug for the URL and an English name.

That's the downside of an app structure that focused on being a CLI first, before realizing it's true power
of being an easy API abstraction.

Maybe one day I'll re-design this thing to be more robust, with an API-first approach, but this is not going to be today.